requirement.txt
discord.js==1.0.1
